re-organize wpp explorer code
